requirements.txt
-i https://pypi.org/simple
immudb-py==1.0.5.post1
